#17 Move XML handling for `OverpassTag` to a dedicated extension file
This will allow for the use of the model without depending on the XML parsing.
