+++ SwiftOverpass/OverpassTag.swift
@@ -7,7 +7,6 @@
 //
 
 import Foundation
-import AEXML
 
 /// Represents a tag for the query.
 public struct OverpassTag {
@@ -32,27 +31,5 @@ public struct OverpassTag {
         self.value = value
         self.isNegation = isNegation
         self.isRegex = isRegex
-    }
-    
-    // MARK: - Internal Functions
-    
-    /**
-     Makes a <has-kv> element
-     */
-    internal func makeHasKvElement() -> AEXMLElement {
-        let aValue = value ?? ""
-        var attributes = ["k" : key]
-        
-        if isNegation {
-            attributes["modv"] = "not"
-        }
-        
-        if isRegex {
-            attributes["regv"] = aValue
-        } else {
-            attributes["v"] = aValue
-        }
-        
-        return AEXMLElement(name: "has-kv", attributes: attributes)
     }
 }
